Move weakCache into Agent
