+++ lib/agent.js
@@ -2,14 +2,25 @@
 
 const { InvalidArgumentError, InvalidReturnValueError } = require('./core/errors')
 const Pool = require('./pool')
-const { weakCache } = require('./core/util')
+const { parseOrigin } = require('./core/util')
 
-const kAgentFactory = Symbol('agent factory')
+const kOpts = Symbol('opts')
+const kCache = Symbol('cache')
+const kCleanup = Symbol('cleanup')
 
 class Agent {
   constructor (opts) {
-    this[kAgentFactory] = weakCache(origin => new Pool(origin, opts))
-    this.origins = []
+    this[kOpts] = opts
+    this[kCache] = new Map()
+    this[kCleanup] = new FinalizationRegistry(key => {
+      // get the WeakRef from the cache
+      const ref = cache.get(key)
+      // if the WeakRef exists and the object has been reclaimed
+      if (ref !== undefined && ref.deref() === undefined) {
+        // remove the WeakRef from the cache
+        cache.delete(key)
+      }
+    })
   }
 
   get (origin) {
@@ -17,27 +28,47 @@ class Agent {
       throw new InvalidArgumentError('Origin must be a non-empty string.')
     }
 
-    if (this.origins.indexOf(origin) === -1) {
-      this.origins.push(origin)
+    // check the cache for an existing WeakRef
+    const ref = this[kCache].get(origin)
+
+    // if one exists in the cache try to return the WeakRef
+    if (ref !== undefined) {
+      const cached = ref.deref()
+      if (cached !== undefined) {
+        return cached
+      }
     }
 
-    return this[kAgentFactory](origin)
+    // otherwise, if it isn't in the cache or the reference has been cleaned up, create a new one!
+    const value = new Pool(origin, this[kOpts])
+    // add a WeakRef of the value to the cache
+    this[kCache].set(origin, new WeakRef(value))
+    // add the value to the finalization registry
+    this[kCleanup].register(value, origin)
+
+    return value
   }
 
  close () {
     const closePromises = []
-    for (const origin in origins) {
-      const client = this[kAgentFactory](origin)
-      closePromises.push(client.close.bind(client)())
+    for (const ref in this[kCache].values()) {
+      const pool = ref.deref()
+
+      if (pool) {
+        closePromises.push(pool.close())
+      }
     }
     return Promise.all(closePromises)
   }
 
- destroy () {
+  destroy () {
     const destroyPromises = []
-    for (const origin in origins) {
-      const client = this[kAgentFactory](origin)
-      destroyPromises.push(client.destroy.bind(client)())
+    for (const ref in this[kCache].values()) {
+      const pool = ref.deref()
+
+      if (pool) {
+        destroyPromises.push(pool.destroy())
+      }
     }
     return Promise.all(destroyPromises)
   }
@@ -58,24 +89,11 @@ function setGlobalAgent (agent) {
 }
 
 function dispatchFromAgent (requestType) {
-  return (url, { agent = globalAgent, method = 'GET', ...opts } = {}, ...additionalArgs) => {
-    if (url === undefined || url === null) {
-      throw new InvalidArgumentError('Argument url must be defined')
-    }
+  return (origin, { agent = globalAgent, method = 'GET', ...opts } = {}, ...additionalArgs) => {
+    const url = parseOrigin(origin)
 
     if (!agent || typeof agent.get !== 'function') {
       throw new InvalidArgumentError('Argument agent must implement Agent')
-    }
-
-    if (opts.path != null) {
-      throw new InvalidArgumentError('unsupported opts.path')
-    }
-
-    // if url is a string transform into URL inst. otherwise use as is
-    url = typeof url === 'string' ? new URL(url) : url
-    // ensure it atleast has a non-empty string origin
-    if (!url || typeof url.origin !== 'string' || url.origin === '') {
-      throw new InvalidArgumentError('Argument url.origin must be a non-empty string')
     }
 
     const client = agent.get(url.origin)
@@ -84,9 +102,7 @@ function dispatchFromAgent (requestType) {
       throw new InvalidReturnValueError(`Client returned from Agent.get() does not implement method ${requestType}`)
     }
 
-    const path = url.path || `${url.pathname || '/'}${url.search || ''}`
-
-    return client[requestType]({ ...opts, method, path }, ...additionalArgs)
+    return client[requestType]({ ...opts, method, path: url.path }, ...additionalArgs)
   }
 }
