+++ lib/core/util.js
@@ -148,45 +148,6 @@ function errnoException (code, syscall) {
   return err
 }
 
-/* istanbul ignore next: https://github.com/tc39/proposal-weakrefs */
-function weakCache (fn) {
-  /* istanbul ignore next: */
-  if (typeof WeakRef === 'undefined' || typeof FinalizationRegistry === 'undefined') {
-    throw new NotSupportedError('In order to use this feature, `WeakRef` and `FinalizationRegistry` must be defined as global objects. Check your Node.js version to be sure it is v14.6.0 or greater.')
-  }
-
-  const cache = new Map()
-  const cleanup = new FinalizationRegistry(key => {
-    // get the WeakRef from the cache
-    const ref = cache.get(key)
-    // if the WeakRef exists and the object has been reclaimed
-    if (ref !== undefined && ref.deref() === undefined) {
-      // remove the WeakRef from the cache
-      cache.delete(key)
-    }
-  })
-  return key => {
-    // check the cache for an existing WeakRef
-    const ref = cache.get(key)
-
-    // if one exists in the cache try to return the WeakRef
-    if (ref !== undefined) {
-      const cached = ref.deref()
-      if (cached !== undefined) {
-        return cached
-      }
-    }
-
-    // otherwise, if it isn't in the cache or the reference has been cleaned up, create a new one!
-    const value = fn(key)
-    // add a WeakRef of the value to the cache
-    cache.set(key, new WeakRef(value))
-    // add the value to the finalization registry
-    cleanup.register(value, key)
-    return value
-  }
-}
-
 module.exports = {
   nop,
   parseOrigin,
@@ -199,5 +160,4 @@ module.exports = {
   destroy,
   bodyLength,
   isBuffer,
-  weakCache
 }
